Refactor CheckoutPage
