+++ src/components/Checkout/CheckoutForm/CheckoutForm.styled.jsx
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+export const StyledForm = styled.form`
+  flex-grow: 1;
+
+  div {
+    margin-block: 1rem;
+  }
+
+  > div:last-of-type {
+    margin-top: 3rem;
+    text-align: right;
+  }
+`;
