+++ src/components/Checkout/CheckoutForm/OrderSummary/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./OrderSummary";
