+++ src/components/Checkout/CheckoutForm/PaymentOptions/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./PaymentOptions";
