+++ src/components/Checkout/CheckoutForm/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./CheckoutForm";
