+++ src/pages/CheckoutPage/CheckoutPage.jsx
@@ -1,0 +1,7 @@
+import Checkout from "../../components/Checkout";
+
+const CheckoutPage = () => {
+  return <Checkout />;
+};
+
+export default CheckoutPage;
